requirements.txt
requests==2.22.0
parse==1.14.0
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.13.0
Autologging==1.3.2
urllib3==1.25.6
uuid==1.30
certifi==2019.9.11
chardet==3.0.4
idna==2.8
scapy==2.4.5
